inventory/requirements.txt
python=3.6.3
Django==2.1.4
djangorestframework==3.9.0